requirements.txt
Flask==3.0.3
Flask-WTF==1.2.2
email_validator==2.2.0
pillow==11.0.0
qrcode==8.0
WTForms==3.2.1